ISHMUMMind_Bloom-AI_assistance_For_Postpartum_Depression-main/requirements.txt
fastapi
uvicorn
pydantic
numpy
pandas
joblib
requests
python-dotenv
scikit-learn
